// src/backend/apps/api/nrw/nrw.d.ts
export type MAENGEL = {
    service_request_id: string;
    status: string;
    status_notes: string;
    service_name: string;
    service_code: string;
    description: string;
    requested_datetime: string;
    updated_datetime: string;
    address: string;
    zipcode: string;
    lat: number;
    long: number;
    media_url: string;
};